Major revision:   changed table attribute to law (and in json too)   removed eigenvalue attribute from an element
Added Elements, Hash, IsSymmetric, HasEvaluation, Evaluation, Eigenvalues and Grading (as a synonym for FinestAdequateGrading) for FusLaw.

Added multiplication for subsets.

Added IsUnit, IsAnnihilating, IsAbsorbing for FusLawElts.

Changed the Grading and Evaluation to be maps from the FusLaw, rather than the set.

Fixed SubConstructor to generate a sublaw rather than take an intersection.

Fixed equality of FusLaws to check for evaluation map.

Fixed printing to do name and evaluation prettily.

Changed AssignEvaluationMap to AssignEvaluation.